PR OT139-28-Componente-Slider : comments fix
